Added Conditionals
Added Conditionals so that when your mouse is neat the circle, it gets smaller and harder to hit and when you're not close enough to the circle, it grow bigger. I also added conditionals so that when the circle gets to 20 or 300, the growing propreties are inverted. This way, the circle never gets too small or to big for us to see clearly.