Looks (WiP) & Gun (WiP)
Changed the look (Work in Progress):
Added a background and set color for each side. When the game is over, the score and winner are relative to the winner.

Added Gun Class (Work in Progress)
they can shoot, but nothing happens yet if you hit the opponent. Plus I'm having issues with the gun returning to the center of the paddle after being used. Also trying to find a way to make the bullet go straight when fired.