Done (not really)
*Works better with flashlight*
Tried to make a little game where you control the wolf with your flashlight. Sheeps are running around and your goal is to kill them all (by getting the wolf over them). Tried to make so the sheep display changes when it dies but couldn't manage to keep the sheepDead display...