Error
Hi, sorry I had once again issues with GitHub and my commits. Everything sould be back to normal now. I made a lot of changes. Mainly, I added a state system with enum. I was planning on imputing a wave system with the program itself but I couldn't manage to find a was so instead Im doing it with the state system... I know doing it this way makes the program longer but I coun't find another way to fo it.